requirements.txt
langchain
langchain-community
langchain-core
langchain-aws
streamlit
python-dotenv
ollama
langsmith
textblob